Create DirectDraw 7 surfaces.
Why DirectDraw, if it's oh so deprecated and sucks and doesn't work?
Well, it works if you use if correctly, and stay away from 8-bit color modes.

There's also that problem with ddraw.lib being missing in some versions of Visual Studio, because those M$ idiots thought it's okay to remove libs just like that.
Oh, well. Maybe we'll need to use LoadLibrary for ddraw.dll, but for now it's the lib.

So, what were the alternatives? I choose not to use SDL, which I got rid of in one of last commits. (or anything else like SDL for that matter)
In Windows there's GDI, which I don't really understand and don't want to.
GDI+ is cool, but not really supported in C (vs C++).
Direct2D is out of the question, I want to keep WinXP compatibility.
And I didn't have time to learn Direct3D.
OpenGL? maybe, but I don't want GL around in non-GL WinQuake.

DirectDraw, then. At least for now.