Now it is possible to comment out fullscreen support.
If `vid_default = MODE_FULLSCREEN_DEFAULT`, is not executed, the game will not set the desired video mode, even if it is windowed.

I am moving that line from VID_InitMGLFull() to VID_Init(), so it is now possible to comment out the call to VID_InitMGLFull() and play with windowed modes only.
The plan is to remove any MGL dependency from windowed modes first, with no fullscreen modes support at all.